pyproject.toml
[project]
name = "chess-agent"
version = "0.1.0"
description = "Agent acting as a chess expert"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "google-adk>=0.1.0",
    "requests>=2.32.3",
]
